requirements.txt
streamlit	
langchain-groq	
langchain-community	
python-dotenv	
pypdf	
chromadb	
yt-dlp	
sentence-transformers	
langchain-huggingface	